utils/requirements.txt
openpyxl==3.1.2  # For handling Excel files
python-docx==0.8.11  # For working with Word documents